# src/server/api/go/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Luminox.",
        "title": "Luminox API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/user/ls": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all users under a project. If limit is not provided or 0, all users will be returned.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "List users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit of users to return. Max 200. If limit is 0 or not provided, all users will be returned.",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cursor for pagination. Use the cursor from the previous response to get the next page.",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": false,
                        "description": "Order by created_at descending if true, ascending if false (default false)",
                        "name": "time_desc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/serializer.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.ListUsersOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "x-code-samples": [
                    {
                        "label": "Python",
                        "lang": "python",
                        "source": "from luminox import LuminoxClient\n\nclient = LuminoxClient(api_key='sk_project_token')\n\n# List all users\nusers = client.users.list()\nfor user in users.items:\n    print(f\"{user.identifier}: {user.id}\")\n\n# List users with pagination\nusers = client.users.list(limit=20, time_desc=True)\nif users.has_more:\n    next_users = client.users.list(limit=20, cursor=users.next_cursor)\n"
                    },
                    {
                        "label": "JavaScript",
                        "lang": "javascript",
                        "source": "import { LuminoxClient } from '@luminox/luminox';\n\nconst client = new LuminoxClient({ apiKey: 'sk_project_token' });\n\n// List all users\nconst users = await client.users.list();\nfor (const user of users.items) {\n  console.log(`${user.identifier}: ${user.id}`);\n}\n\n// List users with pagination\nconst paginatedUsers = await client.users.list({ limit: 20, timeDesc: true });\nif (paginatedUsers.hasMore) {\n  const nextUsers = await client.users.list({ limit: 20, cursor: paginatedUsers.nextCursor });\n}\n"
                    }
                ]
            }
        },
        "/user/{identifier}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a user by identifier and cascade delete all associated resources (Space, Session, Disk, Skill)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Delete user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User identifier string",
                        "name": "identifier",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                },
                "x-code-samples": [
                    {
                        "label": "Python",
                        "lang": "python",
                        "source": "from luminox import LuminoxClient\n\nclient = LuminoxClient(api_key='sk_project_token')\n\n# Delete a user and all associated resources\nclient.users.delete('alice@luminox.io')\n"
                    },
                    {
                        "label": "JavaScript",
                        "lang": "javascript",
                        "source": "import { LuminoxClient } from '@luminox/luminox';\n\nconst client = new LuminoxClient({ apiKey: 'sk_project_token' });\n\n// Delete a user and all associated resources\nawait client.users.delete('alice@luminox.io');\n"
                    }
                ]
            }
        },
        "/user/{identifier}/resources": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the resource counts (Spaces, Sessions, Disks, Skills) associated with a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user resources",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User identifier string",
                        "name": "identifier",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/serializer.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.GetUserResourcesOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "x-code-samples": [
                    {
                        "label": "Python",
                        "lang": "python",
                        "source": "from luminox import LuminoxClient\n\nclient = LuminoxClient(api_key='sk_project_token')\n\n# Get user resource counts\nresources = client.users.get_resources('alice@luminox.io')\nprint(f\"Spaces: {resources.spaces_count}\")\nprint(f\"Sessions: {resources.sessions_count}\")\nprint(f\"Disks: {resources.disks_count}\")\nprint(f\"Skills: {resources.skills_count}\")\n"
                    },
                    {
                        "label": "JavaScript",
                        "lang": "javascript",
                        "source": "import { LuminoxClient } from '@luminox/luminox';\n\nconst client = new LuminoxClient({ apiKey: 'sk_project_token' });\n\n// Get user resource counts\nconst resources = await client.users.getResources('alice@luminox.io');\nconsole.log(`Spaces: ${resources.spaces_count}`);\nconsole.log(`Sessions: ${resources.sessions_count}`);\nconsole.log(`Disks: ${resources.disks_count}`);\nconsole.log(`Skills: ${resources.skills_count}`);\n"
                    }
                ]
            }
        }
    },
    "definitions": {
...
        "model.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "identifier": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "repo.UserResourceCounts": {
            "type": "object",
            "properties": {
                "disks_count": {
                    "type": "integer"
                },
                "sessions_count": {
                    "type": "integer"
                },
                "skills_count": {
                    "type": "integer"
                },
                "spaces_count": {
                    "type": "integer"
                }
            }
        },
        "serializer.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "service.GetUserResourcesOutput": {
            "type": "object",
            "properties": {
                "counts": {
                    "$ref": "#/definitions/repo.UserResourceCounts"
                }
            }
        },
...
        "service.ListUsersOutput": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Project Bearer token (e.g., \"Bearer sk-ac-xxxx\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}